Basic styling and instructions functionality
Finished styling for the basic layout of the game, as well as adding
an instructions modal for increased user experience. When the page is
loaded, the user will be able to read the instructions, and then click
on the instructions button at any point in time to reread them as
neccessary. Progress has been a little slow on this project as I just
drove across the country.