Made some improvements
Noticed when I was deploying my shopping cart project that there were
some accessibility styling problems with this project so I fixed those.
While I was doing that, I noticed that it would be nice to have a modal
pop up to let the user know when they lost (not just when they won), in
case they weren't watching the score in the top right. Overall, took
about 15 minutes and greatly improved the project.